Split cloudlinux actors into multiple single-function
