+++ repos/system_upgrade/cloudlinux/actors/enablerpmnewfiles/libraries/enablerpmnewfiles.py
@@ -21,10 +21,3 @@ def rename_rpmnew():
 
             os.rename(base_path, old_path)
             os.rename(new_path, base_path)
-
-def add_custom():
-    custom_repo_dir = api.get_common_folder_path(CUSTOM_REPOS_FOLDER)
-
-    for repofile in os.listdir(custom_repo_dir):
-        full_repo_path = os.path.join(custom_repo_dir, repofile)
-        shutil.copy(full_repo_path, REPO_ROOT_PATH)
